progress: extract progress updating into Updater struct
This allows reusing the code to create periodic progress updates.
