+++ internal/ui/progress/counter_test.go
@@ -35,7 +35,7 @@ func TestCounter(t *testing.T) {
 		lastTotal = total
 		ncalls++
 	}
-	c := progress.New(10*time.Millisecond, startTotal, report)
+	c := progress.NewCounter(10*time.Millisecond, startTotal, report)
 
 	done := make(chan struct{})
 	go func() {
@@ -63,24 +63,6 @@ func TestCounterNil(t *testing.T) {
 	// Shouldn't panic.
 	var c *progress.Counter
 	c.Add(1)
+	c.SetMax(42)
 	c.Done()
-}
-
-func TestCounterNoTick(t *testing.T) {
-	finalSeen := false
-	otherSeen := false
-
-	report := func(value, total uint64, d time.Duration, final bool) {
-		if final {
-			finalSeen = true
-		} else {
-			otherSeen = true
-		}
-	}
-	c := progress.New(0, 1, report)
-	time.Sleep(time.Millisecond)
-	c.Done()
-
-	test.Assert(t, finalSeen, "final call did not happen")
-	test.Assert(t, !otherSeen, "unexpected status update")
 }
